feat(backend): add Proveedor and Insumo models, support fractional quantities

+++ backend/app/api/v1/endpoints/desarrollo.py
@@ -11,7 +11,7 @@ router = APIRouter()
 
 
 class ItemPedidoMaterial(BaseModel):
-    cantidad: int = Field(gt=0)
+    cantidad: float = Field(gt=0)
     descripcion: str
     uso_en: str = ""
     observaciones: str = ""
